+++ src/mc_classifier_pipeline/__init__.py
@@ -2,19 +2,8 @@
 
 __version__ = "0.0.1"
 
-from mc_classifier_pipeline import (
-    bert_recipe,
-    doc_retriever,
-    evaluation,
-    label_studio_uploader,
-    preprocessing,
-    run_pipeline,
-    sk_naive_bayes_recipe,
-    trainer,
-    utils,
-)
-
 __all__ = [
+    "annotation_analysis",
     "bert_recipe",
     "doc_retriever",
     "evaluation",
@@ -24,5 +13,6 @@ __all__ = [
     "sk_naive_bayes_recipe",
     "trainer",
     "utils",
+    "model_orchestrator",
     "__version__",
 ]